Split imap/connection.go into two separate files: distributor/connection.go for "real", outgoing connections, and imap/session.go for internal, proxied connection tracking.
